+++ src/main/java/com/github/unchama/buildassist/listener/PlayerJoinListener.java
@@ -1,6 +1,8 @@
 /*** Eclipse Class Decompiler plugin, copyright (c) 2012 Chao Chen (cnfree2000@hotmail.com) ***/
-package com.github.unchama.buildassist;
+package com.github.unchama.buildassist.listener;
 
+import com.github.unchama.buildassist.BuildAssist;
+import com.github.unchama.buildassist.LoadPlayerDataTaskRunnable;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
