+++ src/main/java/com/github/unchama/buildassist/listener/PlayerQuitListener.java
@@ -1,5 +1,7 @@
-package com.github.unchama.buildassist;
+package com.github.unchama.buildassist.listener;
 
+import com.github.unchama.buildassist.BuildAssist;
+import com.github.unchama.buildassist.PlayerData;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
