+++ src/main/java/com/github/unchama/buildassist/listener/TypedEventListener.kt
@@ -1,0 +1,12 @@
+package com.github.unchama.buildassist.listener
+
+import org.bukkit.event.EventHandler
+import org.bukkit.event.Listener
+
+/**
+ * ラムダ用通常優先度のリスナー
+ */
+interface TypedEventListener<E> : Listener {
+  @EventHandler
+  fun onEvent(event: E)
+}
